+++ beam-migrate/Database/Beam/Migrate/Checks.hs
@@ -22,6 +22,21 @@ import Data.Semigroup
 
 import GHC.Generics (Generic)
 
+-- * Schema checks
+
+-- | Asserts that a schema with the given name exists in a database
+data SchemaExistsPredicate = SchemaExistsPredicate Text {-^ Table name -}
+  deriving (Show, Eq, Ord, Typeable, Generic)
+instance Hashable SchemaExistsPredicate
+instance DatabasePredicate SchemaExistsPredicate where
+  englishDescription (SchemaExistsPredicate s) =
+    "Schema " <> show s <> " must exist"
+
+  serializePredicate (SchemaExistsPredicate s) =
+    object [ "schema-exists" .= s ]
+
+  predicateSpecificity _ = PredicateSpecificityAllBackends
+
 -- * Table checks
 
 -- | Asserts that a table with the given name exists in a database
@@ -131,12 +146,18 @@ beamCheckDeserializers
      , HasDataTypeCreatedCheck (BeamMigrateSqlBackendDataTypeSyntax be) )
   => BeamDeserializers be
 beamCheckDeserializers = mconcat
-  [ beamDeserializer (const deserializeTableExistsPredicate)
+  [ beamDeserializer (const deserializeSchemaExistsPredicate)
+  , beamDeserializer (const deserializeTableExistsPredicate)
   , beamDeserializer (const deserializeTableHasPrimaryKeyPredicate)
   , beamDeserializer deserializeTableHasColumnPredicate
   , beamDeserializer deserializeTableColumnHasConstraintPredicate
   ]
   where
+    deserializeSchemaExistsPredicate :: Value -> Parser SomeDatabasePredicate
+    deserializeSchemaExistsPredicate =
+      withObject "SchemaExistsPredicate" $ \v ->
+      SomeDatabasePredicate <$> (SchemaExistsPredicate <$> v .: "schema-exists")
+
     deserializeTableExistsPredicate :: Value -> Parser SomeDatabasePredicate
     deserializeTableExistsPredicate =
       withObject "TableExistPredicate" $ \v ->