+++ beam-postgres/test/Database/Beam/Postgres/Test/Migrate.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE LambdaCase #-}
 module Database.Beam.Postgres.Test.Migrate where
 
 import Database.Beam
@@ -22,6 +23,8 @@ tests postgresConn =
       , charWidthVerification postgresConn "CHAR" char
       , charNoWidthVerification postgresConn "CHAR" char
       , extensionVerification postgresConn
+      , createTableWithSchemaWorks postgresConn
+      , dropSchemaWorks postgresConn
       ]
 
 data CharT f
@@ -96,3 +99,41 @@ extensionVerification pgConn =
           case resAfter of
             VerificationSucceeded -> return ()
             VerificationFailed failures -> fail ("Verification failed: " ++ show failures)
+
+
+-- | Verifies that 'createTableWithSchema' correctly creates a table
+-- with a schema.
+createTableWithSchemaWorks :: IO ByteString -> TestTree
+createTableWithSchemaWorks pgConn =
+    testCase ("createTableWithSchema works correctly") $ do
+      withTestPostgres "create_table_with_schema" pgConn $ \conn -> do
+        res <- runBeamPostgres conn $ do
+          db <- executeMigration runNoReturn $ do
+                  internalSchema <- createDatabaseSchema "internal_schema"
+                  (CharDb <$> createTableWithSchema (Just internalSchema) "char_test"
+                                    (CharT (field "key" (varchar Nothing) notNull)))
+
+          verifySchema migrationBackend db
+
+        case res of
+          VerificationSucceeded -> return ()
+          VerificationFailed failures -> fail ("Verification failed: " ++ show failures)
+
+
+-- | Verifies that creating a schema and dropping it works
+dropSchemaWorks :: IO ByteString -> TestTree
+dropSchemaWorks pgConn =
+    testCase ("dropDatabaseSchema works correctly") $ do
+      withTestPostgres "drop_schema" pgConn $ \conn -> do
+        runBeamPostgres conn $ do
+          db <- executeMigration runNoReturn $ do
+                  internalSchema <- createDatabaseSchema "internal_schema"
+                  willBeDroppedSchema <- createDatabaseSchema "will_be_dropped"
+                  db <- (CharDb <$> createTableWithSchema (Just internalSchema) "char_test"
+                                    (CharT (field "key" (varchar Nothing) notNull)))
+                  dropDatabaseSchema willBeDroppedSchema
+                  pure db
+
+          verifySchema migrationBackend db >>= \case
+            VerificationFailed failures -> fail ("Verification failed: " ++ show failures)
+            VerificationSucceeded -> pure ()