+++ beam-postgres/ChangeLog.md
@@ -1,8 +1,9 @@
-# Unreleased
+# 0.5.4.0
 
 ## Added features
 
  * Better error messages on column type mismatches (#696).
+ * Added support for creating and dropping database schemas and associated tables with `createDatabaseSchema`, `dropDatabaseSchema`, and `createTableWithSchema` (#716).
 
 ## Documentation